Refact: type 분리
