+++ src/components/Modal/Intro.tsx
@@ -7,18 +7,7 @@ import { auth } from "../../../firebaseConfig";
 import { doc, getDoc, updateDoc } from "firebase/firestore";
 import styled from "styled-components";
 import ProfileAddIcon from "../../assets/icons/ProfileAddIcon.jpeg";
-
-interface NewIntro {
-  imageUrl: string;
-  NickName: string;
-}
-
-interface ChangeIntroProps {
-  closeModal: () => void;
-  changeIntro: React.Dispatch<React.SetStateAction<NewIntro | null>>;
-  setLoading: React.Dispatch<React.SetStateAction<boolean>>;
-  showToast: (message: string) => void;
-}
+import { ChangeIntroProps } from "../../types/Intro";
 
 function Intro({
   closeModal,
