+++ src/types/List.ts
@@ -1,0 +1,8 @@
+export interface Book {
+  id: string;
+  imageUrl: string;
+  bookTitle: string;
+  author: string;
+  publisher: string;
+  price: string;
+}
