+++ src/types/MyPage.ts
@@ -1,0 +1,4 @@
+export interface NewIntro {
+  imageUrl: string;
+  NickName: string;
+}
