+++ src/types/Review.ts
@@ -1,0 +1,15 @@
+export interface ReviewProps {
+  bookId: string;
+  bookTitle: string;
+  author: string;
+  publisher: string;
+  price: string;
+  imageUrl: string;
+  onDelete: (id: string) => void;
+  onClose: () => void;
+}
+
+export interface Comment {
+  NickName: string;
+  comment: string;
+}
